Extended return code handling for msi and exe installers. http://dotnetinstaller.codeplex.com/WorkItem/View.aspx?WorkItemId=3924
Added returncodes_reboot to cmd components that specify which return code(s) should be treated as a required reboot. Added returncodes_failure to cmd components that specify a list of comma-separated return code(s) that should be treated as a failure. Specify "none" to create a component that doesn't fail based on its return code and specify !value to treat all return codes except a single value as failure.